Get handlers lazily
Currently, getHandler is called before constructing a HandlerInfo object. This
results in work being done upfront even if the `handler` property of the
HandlerInfo is never accessed. This change invokes `getHandler` only when the
value is actually needed.

This is particularly needed in the case of `getHandler` fetching lazily-loaded
routes. Previously lazy routes would be fetched even if the router was just
generating a URL; a task which can be done without the whole handler.